h flag for o fixed to unsigned short

+++ get_mod_arg.c
@@ -6,7 +6,7 @@
 /*   By: angkim <angkim@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/08/26 15:29:47 by angkim            #+#    #+#             */
-/*   Updated: 2019/10/03 15:23:14 by angkim           ###   ########.fr       */
+/*   Updated: 2019/10/06 13:36:51 by angkim           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,7 +35,7 @@ void	get_mod_arg_u(t_format *f, va_list args)
 	else if (f->mod & M_LL)
 		f->u_arg = (unsigned long long)va_arg(args, unsigned long long);
 	else if (f->mod & M_H)
-		f->u_arg = (unsigned int)va_arg(args, unsigned int);
+		f->u_arg = (unsigned short)va_arg(args, unsigned int);
 	else if (f->mod & M_HH)
 		f->u_arg = (unsigned char)va_arg(args, int);
 }
@@ -49,7 +49,7 @@ void	get_mod_arg_ox_xx(t_format *f, va_list args)
 	else if (f->mod & M_LL)
 		f->ox_arg = (unsigned long long)va_arg(args, unsigned long long);
 	else if (f->mod & M_H)
-		f->ox_arg = (unsigned int)va_arg(args, unsigned int);
+		f->ox_arg = (unsigned short)va_arg(args, unsigned int);
 	else if (f->mod & M_HH)
 		f->ox_arg = (unsigned char)va_arg(args, unsigned int);
 }
